move deleteSearch function to separate file
